requirements.txt
# Symbolica Benchmark Dependencies
symbolica>=0.1.0
openai>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
pytest>=7.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0